+++ pkg/vault/vault_test.sh
@@ -18,7 +18,7 @@ test() {
   # Get Vault's root token
   export VAULT_TOKEN=$(kubectl get secrets vault-unseal-keys -n $KUBE_NAMESPACE -o jsonpath={.data.vault-root} | base64 --decode)
   export VAULT_ADDRESS="http://127.0.0.1:$PORT_TO_FORWARD/"
-  go test -v vault_interface.go vault_interface_test.go
+  go test -v vault_interface.go vault_impl.go vault_interface_test.go
   # Kill the port-forward if nessecarry 
   kill -9 %%
 }
